Issue#1733 city confirmation (#1734)
* Fixed incorrectly working geolocation confirmation

* Fixed saving geolocation to local storage before user confirmation

* Added change of geolocation to standard (Kyiv) when clicking on the  "Change" button

* Added focus to the search input when clicking on the "Change" button

* Fixed cropped cross icon

* Removed wrapping of additional text about the city

* Reduced the width of the sliced text

* Fixed saving geolocation when location is forbidden

* Added a condition when Kyiv is already in the search, and we only need to call the list of available places without a request

+++ src/app/shared/store/filter.state.ts
@@ -147,10 +147,12 @@ export class FilterState {
   constructor(private appWorkshopsService: AppWorkshopsService) {}
 
   @Action(SetCity)
-  setCity({ patchState, dispatch, getState }: StateContext<FilterStateModel>, { payload }: SetCity): void {
+  setCity({ patchState, dispatch, getState }: StateContext<FilterStateModel>, { payload, isConfirmedCity }: SetCity): void {
     patchState({ settlement: payload });
-    localStorage.setItem('cityConfirmation', JSON.stringify(payload));
-
+    if (isConfirmedCity) {
+      localStorage.setItem('cityConfirmation', JSON.stringify(payload));
+    }
+    
     dispatch(new FilterChange());
   }
